Create README.md
Hello everyone,

I am Anish Kumar. I am working on a project to build AI based social media website for creators and visual artists. The core of my project is Humans and creativity. With the advent and advancements in AI, i believe we are in the next stage of computation revolution. It would be a crime not to be a part of it as its now or never. 

I am looking to work with people having similar mindset but with good tech skills who could help me in this journey. I am cool with just guiding too on the tech part. I am reachable here or can be reached on my email at anishk2940@gmail.com.